requirements.txt
django-mptt>=0.5.2
